requirements.txt
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Limiter==4.0.0
Werkzeug==3.1.5
Jinja2==3.1.6
APScheduler==3.11.2
SQLAlchemy==2.0.45
requests==2.32.5
python-dotenv==1.2.1
openai==2.15.0
httpx==0.28.1
msal==1.34.0
psycopg2-binary==2.9.11
PyMySQL==1.1.0
gunicorn==23.0.0
email-validator==2.2.0
itsdangerous==2.2.0
PyJWT==2.10.1
WTForms==3.2.1
twilio>=8.0.0
beautifulsoup4==4.12.3
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
